+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.11.0] - 2026-02-09
+
+### Added
+
+- **Browser Themes** info module under UI -- Catppuccin Mocha install guide for Brave, Chromium, Firefox and LibreWolf
+
+### Changed
+
+- Renamed GNOME module to **UI** in main menu, health checks and breadcrumbs
+
 ## [0.10.9] - 2026-02-09
 
 ### Fixed